move useBrowserStorage hook to a separate file
