+++ app/packages/state/src/hooks/index.ts
@@ -1,9 +1,10 @@
 export * from "./hooks-utils";
 export {
-  default as useBeforeScreenshot,
   BeforeScreenshotContext,
   callbacks as screenshotCallbacks,
+  default as useBeforeScreenshot,
 } from "./useBeforeScreenshot";
+export * from "./useBrowserStorage";
 export { default as useClearModal } from "./useClearModal";
 export { default as useCreateLooker } from "./useCreateLooker";
 export { default as useDimensions } from "./useDimensions";
@@ -12,36 +13,36 @@ export { default as useHelpPanel } from "./useHelpPanel";
 export { default as useHover } from "./useHover";
 export { default as useHoveredSample } from "./useHoveredSample";
 export { default as useJSONPanel } from "./useJSONPanel";
-export { default as useLookerStore } from "./useLookerStore";
 export * from "./useLookerStore";
+export { default as useLookerStore } from "./useLookerStore";
+export * from "./useOnSelectLabel";
 export { default as usePanel } from "./usePanel";
 export { default as useReset } from "./useReset";
-export { default as useSavedViews } from "./useSavedViews";
-export { default as useScreenshot } from "./useScreenshot";
 export {
   default as useRouter,
   getCurrentEnvironment,
   setCurrentEnvironment,
 } from "./useRouter";
+export { default as useSavedViews } from "./useSavedViews";
+export { default as useScreenshot } from "./useScreenshot";
 export { default as useSelectFlashlightSample } from "./useSelectFlashlightSample";
 export { default as useSelectSample } from "./useSelectSample";
 export { default as useSendEvent } from "./useSendEvent";
+export { default as useSessionSpaces } from "./useSessionSpaces";
 export { default as useSetDataset } from "./useSetDataset";
 export { default as useSetExpandedSample } from "./useSetExpandedSample";
+export { default as useSetExtendedSelection } from "./useSetExtendedSelection";
 export { default as useSetGroupSlice } from "./useSetGroupSlice";
 export { default as useSetSelected } from "./useSetSelected";
 export { default as useSetSelectedLabels } from "./useSetSelectedLabels";
+export { default as useSetSpaces } from "./useSetSpaces";
 export { default as useSetView, stateProxy } from "./useSetView";
+export * from "./useStateUpdate";
 export { default as useStateUpdate } from "./useStateUpdate";
-export * from "./useStateUpdate";
-export { default as useTooltip } from "./useTooltip";
-export * from "./useOnSelectLabel";
-export * from "./useStateUpdate";
 export { default as useTo } from "./useTo";
 export { default as useToClips } from "./useToClips";
 export { default as useToEvaluationPatches } from "./useToEvaluationPatches";
+export { default as useTooltip } from "./useTooltip";
 export { default as useToPatches } from "./useToPatches";
 export { default as useUpdateSample } from "./useUpdateSample";
-export { default as useSetSpaces } from "./useSetSpaces";
-export { default as useSessionSpaces } from "./useSessionSpaces";
 export { default as useResetExtendedSelection } from "./useResetExtendedSelection";
